requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
watchdog
pandas
shutil
asyncio
pydantic